Fluff C1: split lib/api.ts into domain modules behind a re-exporting barrel
Mechanical, behaviour-preserving (audit M2.1). api.ts becomes a pure
barrel re-exporting lib/api/{matters,documents,assistant,modules,
signoffs,audit,admin,settings} plus the existing _core and auth.
Code moved verbatim by domain; export surface verified identical
(213 symbols before and after); zero import changes outside lib/.
