+++ frontend/src/lib/api.ts
@@ -1,2133 +1,17 @@
-// Thin fetch wrapper for the Legalise backend.
-// All endpoints live under /api.
-
-export interface Matter {
-  id: string;
-  slug: string;
-  title: string;
-  matter_type: string;
-  cause: string | null;
-  status: string;
-  case_theory: string | null;
-  pivot_fact: string | null;
-  privilege_posture: string;
-  default_model_id: string;
-  // Keyed provider the default model needs ("anthropic"/"openai"), or
-  // null for keyless models. Backend-supplied truth (provider_for_model);
-  // the FE no longer re-derives model families.
-  required_provider: string | null;
-  facts: Record<string, unknown>;
-  opened_at: string;
-  closed_at: string | null;
-  retention_until: string | null;
-  created_by_id: string;
-}
-
-export interface MatterDocument {
-  id: string;
-  matter_id: string;
-  filename: string;
-  mime_type: string;
-  size_bytes: number;
-  sha256: string;
-  tag: string | null;
-  from_disclosure: boolean;
-  uploaded_at: string;
-  uploaded_by_id: string;
-  comment_count?: number;
-  open_comment_count?: number;
-  version_count?: number;
-  edit_count?: number;
-  pending_edit_count?: number;
-}
-
-export interface MatterCreate {
-  title: string;
-  matter_type?: string;
-  cause?: string | null;
-  case_theory?: string | null;
-  pivot_fact?: string | null;
-  privilege_posture?: string;
-  default_model_id?: string;
-  facts?: Record<string, unknown>;
-  retention_until?: string | null;
-}
-
-// Core fetch wrapper, error envelopes, and origin constants live in
-// `./api/_core`. Re-export so existing `../lib/api` import paths keep
-// resolving every public symbol unchanged.
-export {
-  API,
-  BACKEND_ROOT,
-  ProviderKeyMissingError,
-  ProviderUpstreamError,
-  providerKeyMissingFromBody,
-  providerUpstreamMessage,
-  tryParseProviderUpstream,
-  jsonOrThrow,
-  apiFetch,
-} from "./api/_core";
-export type { ProviderUpstreamCode } from "./api/_core";
-
-// Local imports for everything below that still uses the core helpers.
-import {
-  API,
-  BACKEND_ROOT,
-  jsonOrThrow,
-  apiFetch,
-} from "./api/_core";
-import { AUTH } from "./api/auth";
-
-export interface BootstrapState {
-  user_count: number;
-  has_superuser: boolean;
-  // When false (default), the firm role hierarchy is dormant:
-  // don't present B_mixed qualified-solicitor blockers.
-  firm_role_gates_enabled?: boolean;
-}
-
-// No auth required. The /app first-run screen reads this to decide
-// between empty-state / bootstrap-required / authed-home.
-export const getBootstrapState = () =>
-  apiFetch(`${API}/system/bootstrap-state`).then((r) =>
-    jsonOrThrow<BootstrapState>(r),
-  );
-
-// ---------------------------------------------------------------------------
-// v2 module catalog + trust ceremony
-// ---------------------------------------------------------------------------
-
-export interface V2ManifestEntry {
-  module_id: string;
-  source_kind: string;
-  manifest: Record<string, unknown>;
-  is_valid: boolean;
-  validation_errors: Array<{ path: string; message: string }>;
-}
-
-export interface V2RegistryResponse {
-  modules: V2ManifestEntry[];
-  ui_slots: string[];
-}
-
-export const getModulesV2 = () =>
-  apiFetch(`${API}/modules/v2`).then((r) =>
-    jsonOrThrow<V2RegistryResponse>(r),
-  );
-
-export const getModuleV2 = (moduleId: string) =>
-  apiFetch(`${API}/modules/v2/${encodeURIComponent(moduleId)}`).then((r) =>
-    jsonOrThrow<V2ManifestEntry>(r),
-  );
-
-// Module Standalone + Create Module v1 — read-only manifest validation
-// (same validator as install). Authed; no DB write / ceremony / audit.
-export interface ManifestValidationError {
-  path: string;
-  message: string;
-}
-export interface ValidateManifestResult {
-  valid: boolean;
-  errors: ManifestValidationError[];
-}
-export const validateManifest = (manifest: unknown) =>
-  apiFetch(`${API}/modules/validate`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ manifest }),
-  }).then((r) => jsonOrThrow<ValidateManifestResult>(r));
-
-// Lawve Skill Importer v1 — browse external skills + convert to a
-// governed module draft (read-only; never installs/executes scripts).
-export interface LawveSkillRow {
-  source: string;
-  repo: string;
-  ref: string | null;
-  slug: string;
-  name: string;
-  description: string;
-  version: string | null;
-  author_name: string | null;
-  license: string | null;
-  source_path: string | null;
-  /** Direct attribution link to the skill's lawve.ai directory page. */
-  lawve_url?: string | null;
-  has_references: boolean;
-  has_scripts: boolean;
-  script_review_required: boolean;
-}
-export interface LawveSkillDetail extends LawveSkillRow {
-  skill_markdown: string;
-  frontmatter: Record<string, unknown>;
-  references: string[];
-  scripts: string[];
-  license_text: string | null;
-  provenance: { repo_url: string; ref: string | null; source_path: string | null };
-}
-export interface LawveDraftWarning {
-  code: string;
-  message: string;
-}
-export interface LawveDraftResult {
-  manifest: Record<string, unknown>;
-  valid: boolean;
-  errors: ManifestValidationError[];
-  warnings: LawveDraftWarning[];
-  source_provenance: { repo_url: string; ref: string | null; source_path: string | null } | null;
-  next_steps: string[];
-}
-
-export const listLawveSkills = () =>
-  apiFetch(`${API}/modules/external/lawve/skills`).then((r) =>
-    jsonOrThrow<{ source: string; repo: string; ref: string | null; skills: LawveSkillRow[] }>(r),
-  );
-
-// The honest gap strip — how many skills the lawve.ai directory lists
-// versus how many are importable here today. Public; cached 1h
-// server-side; callers hide the strip on failure.
-export interface LawveDirectoryCount {
-  source: string;
-  skills_url: string;
-  count: number;
-}
-export const getLawveDirectoryCount = () =>
-  apiFetch(`${API}/modules/external/lawve/directory-count`).then((r) =>
-    jsonOrThrow<LawveDirectoryCount>(r),
-  );
-
-export const getLawveSkill = (slug: string) =>
-  apiFetch(`${API}/modules/external/lawve/skills/${encodeURIComponent(slug)}`).then((r) =>
-    jsonOrThrow<LawveSkillDetail>(r),
-  );
-
-export const draftLawveModule = (slug: string, overrides?: Record<string, unknown>) =>
-  apiFetch(`${API}/modules/external/lawve/skills/${encodeURIComponent(slug)}/draft`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(overrides ?? {}),
-  }).then((r) => jsonOrThrow<LawveDraftResult>(r));
-
-export const getGithubSkill = (url: string) =>
-  apiFetch(`${API}/modules/external/github/skill?url=${encodeURIComponent(url)}`).then(
-    (r) => jsonOrThrow<LawveSkillDetail>(r),
-  );
-
-export const draftGithubModule = (url: string, overrides?: Record<string, unknown>) =>
-  apiFetch(`${API}/modules/external/github/draft`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ url, ...(overrides ?? {}) }),
-  }).then((r) => jsonOrThrow<LawveDraftResult>(r));
-
-// Installed-modules listing. One row per module_id
-// (most recent installed_at). Frontend uses it for the catalog
-// badge and as one AND clause in GrantsPanel.runnablePairs.
-export interface InstalledModule {
-  module_id: string;
-  name?: string | null;
-  version: string;
-  publisher: string;
-  visibility: string;
-  signature_status: string;
-  capabilities?: unknown[];
-  enabled: boolean;
-  installed_at: string;
-  installed_by_user_id: string | null;
-  install_path?: string | null;
-  track_record?: Record<string, number>;
-}
-
-export const listInstalledModules = () =>
-  apiFetch(`${API}/modules/installed`).then((r) =>
-    jsonOrThrow<InstalledModule[]>(r),
-  );
-
-// Skill requests. The audit chain is the store — POST writes one
-// module.request.created row; GET (admin) derives the pending set
-// from those rows minus module_ids already installed and enabled.
-export interface ModuleRequestRow {
-  module_id: string;
-  source: string | null;
-  requested_by: string | null;
-  requested_at: string;
-}
-
-export const requestModule = (moduleId: string, source?: string) =>
-  apiFetch(`${API}/modules/requests`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ module_id: moduleId, source: source ?? null }),
-  }).then((r) => jsonOrThrow<{ ok: boolean }>(r));
-
-export const listModuleRequests = () =>
-  apiFetch(`${API}/modules/requests`).then((r) =>
-    jsonOrThrow<ModuleRequestRow[]>(r),
-  );
-
-export interface CeremonyPermissionCard {
-  module_id: string;
-  module_name?: string;
-  publisher?: string;
-  publisher_verified?: boolean;
-  signature_status?: string;
-  visibility?: string;
-  version?: string;
-  capabilities?: unknown[];
-  data_movement_summary?: unknown;
-  gates?: unknown;
-  advice_tier_max?: string;
-  audit_events?: unknown[];
-  dependencies?: unknown[];
-}
-
-export interface CeremonyResponse {
-  ceremony_id: string;
-  module_id: string;
-  state: string;
-  fast_path: boolean;
-  is_terminal: boolean;
-  permission_card: CeremonyPermissionCard;
-  history: Array<Record<string, unknown>>;
-}
-
-export type CeremonyAction = "trust" | "reject" | "grant";
-
-export interface StartInstallRequest {
-  source: "registry" | "manifest";
-  module_id?: string;
-  manifest?: Record<string, unknown>;
-  signature?: string;
-}
-
-export const startInstall = (body: StartInstallRequest) =>
-  apiFetch(`${API}/modules/install`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => jsonOrThrow<CeremonyResponse>(r));
-
-export const getCeremony = (ceremonyId: string) =>
-  apiFetch(`${API}/modules/install/${encodeURIComponent(ceremonyId)}`).then(
-    (r) => jsonOrThrow<CeremonyResponse>(r),
-  );
-
-/**
- * Distinguishable error for the 409 invalid-transition path so the
- * stepper can render a structured banner with the substrate's
- * `module.ceremony.rejected` audit row in mind. The reason + message
- * are populated from the substrate's structured detail body.
- */
-export class InvalidCeremonyTransitionError extends Error {
-  readonly kind = "invalid_ceremony_transition" as const;
-  constructor(
-    message: string,
-    public readonly ceremonyId: string,
-    public readonly requestedAction: CeremonyAction,
-  ) {
-    super(message);
-    this.name = "InvalidCeremonyTransitionError";
-  }
-}
-
-export const advanceCeremony = async (
-  ceremonyId: string,
-  action: CeremonyAction,
-): Promise<CeremonyResponse> => {
-  const res = await apiFetch(
-    `${API}/modules/install/${encodeURIComponent(ceremonyId)}/advance`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ action }),
-    },
-  );
-  if (res.status === 409) {
-    let body: { detail?: { message?: string } } = {};
-    try {
-      body = (await res.json()) as typeof body;
-    } catch {
-      // fall through
-    }
-    throw new InvalidCeremonyTransitionError(
-      body?.detail?.message ?? "Invalid ceremony transition",
-      ceremonyId,
-      action,
-    );
-  }
-  return jsonOrThrow<CeremonyResponse>(res);
-};
-
-export interface UpdateModuleRequest {
-  new_manifest: Record<string, unknown>;
-  signature?: string;
-}
-
-export interface UpdateModuleResponse {
-  module_id: string;
-  new_version: string;
-  expansion_detected: boolean;
-  expansion_report: Record<string, unknown>;
-  ceremony_id: string | null;
-}
-
-export const updateModuleV2 = (moduleId: string, body: UpdateModuleRequest) =>
-  apiFetch(`${API}/modules/${encodeURIComponent(moduleId)}/update`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => jsonOrThrow<UpdateModuleResponse>(r));
-
-export const revokeModuleV2 = (moduleId: string) =>
-  apiFetch(`${API}/modules/${encodeURIComponent(moduleId)}/revoke`, {
-    method: "POST",
-  }).then((r) => jsonOrThrow<{ module_id: string; disabled_rows: number; revoked_grants: number }>(r));
-
-// ---------------------------------------------------------------------------
-// Matter-scoped grants
-// ---------------------------------------------------------------------------
-
-export interface GrantRow {
-  id: string;
-  plugin: string;
-  skill: string;
-  capability: string;
-  scope_type: string;
-  scope_id: string | null;
-  granted_at: string | null;
-}
-
-export interface GrantListResponse {
-  matter_id: string;
-  grants: GrantRow[];
-}
-
-export interface GrantCreateResponse {
-  matter_id: string;
-  parent_capability_id: string;
-  module_id: string;
-  grants: GrantRow[];
-  was_idempotent_noop: boolean;
-}
-
-export const listGrants = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/grants`).then((r) =>
-    jsonOrThrow<GrantListResponse>(r),
-  );
-
-/**
- * Distinguishable errors for the substrate's two non-200 paths on
- * grant creation. The endpoint returns:
- *   - 404 module_not_installed
- *   - 409 module_disabled (installed but admin disabled it)
- * Both carry structured bodies the substrate documents at
- * backend/app/api/grants.py:156-175.
- */
-export class ModuleNotInstalledError extends Error {
-  readonly kind = "module_not_installed" as const;
-  constructor(message: string, public readonly moduleId: string) {
-    super(message);
-    this.name = "ModuleNotInstalledError";
-  }
-}
-
-export class ModuleDisabledError extends Error {
-  readonly kind = "module_disabled" as const;
-  constructor(message: string, public readonly moduleId: string) {
-    super(message);
-    this.name = "ModuleDisabledError";
-  }
-}
-
-export const createGrant = async (
-  slug: string,
-  body: { module_id: string; capability_id: string },
-): Promise<GrantCreateResponse> => {
-  const res = await apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/grants`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(body),
-    },
-  );
-  if (res.status === 404) {
-    let parsed: { detail?: { error?: string; module_id?: string } } = {};
-    try {
-      parsed = (await res.json()) as typeof parsed;
-    } catch {
-      // fall through
-    }
-    if (parsed?.detail?.error === "module_not_installed") {
-      throw new ModuleNotInstalledError(
-        `Skill ${body.module_id} has not been added to this workspace.`,
-        body.module_id,
-      );
-    }
-  }
-  if (res.status === 409) {
-    let parsed: { detail?: { error?: string; message?: string } } = {};
-    try {
-      parsed = (await res.json()) as typeof parsed;
-    } catch {
-      // fall through
-    }
-    if (parsed?.detail?.error === "module_disabled") {
-      throw new ModuleDisabledError(
-        parsed.detail.message ?? "Module is currently disabled.",
-        body.module_id,
-      );
-    }
-  }
-  return jsonOrThrow<GrantCreateResponse>(res);
-};
-
-export const revokeGrant = (slug: string, grantId: string) =>
-  apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/grants/${encodeURIComponent(grantId)}`,
-    { method: "DELETE" },
-  ).then((r) => {
-    if (r.status === 204) return;
-    return jsonOrThrow<unknown>(r).then(() => undefined);
-  });
-
-// ---------------------------------------------------------------------------
-// Invocation + artifacts
-// ---------------------------------------------------------------------------
-
-export interface InvocationResponse {
-  invocation_id: string;
-  module_id: string;
-  capability_id: string;
-  matter_id: string;
-  result: Record<string, unknown>;
-}
-
-/**
- * Substrate truth (backend/app/api/invocations.py): every non-200 path
- * returns a structured body with an `error` discriminator. Surfacing
- * each as a typed error lets the runner branch on `instanceof` rather
- * than parse strings.
- */
-export class PostureBlockedError extends Error {
-  readonly kind = "posture_gate_blocked" as const;
-  constructor(
-    message: string,
-    public readonly posture: string,
-    public readonly requiredRole: string,
-    public readonly actorRole: string,
-    public readonly reason: string,
-  ) {
-    super(message);
-    this.name = "PostureBlockedError";
-  }
-}
-
-export class CapabilityDeniedError extends Error {
-  readonly kind = "capability_denied" as const;
-  constructor(
-    message: string,
-    public readonly plugin: string,
-    public readonly skill: string,
-    public readonly capability: string,
-  ) {
-    super(message);
-    this.name = "CapabilityDeniedError";
-  }
-}
-
-export class Phase1BlockedError extends Error {
-  readonly kind = "phase1_blocked" as const;
-  constructor(
-    message: string,
-    public readonly blockedReason: string,
-    public readonly gateState: Record<string, unknown>,
-  ) {
-    super(message);
-    this.name = "Phase1BlockedError";
-  }
-}
-
-export class ProviderKeyMissingForInvokeError extends Error {
-  readonly kind = "provider_key_missing" as const;
-  constructor(message: string, public readonly provider: string | null) {
-    super(message);
-    this.name = "ProviderKeyMissingForInvokeError";
-  }
-}
-
-export class ProviderUpstreamInvokeError extends Error {
-  readonly kind = "provider_upstream_error" as const;
-  constructor(
-    message: string,
-    public readonly provider: string | null,
-    public readonly code: string | null,
-    public readonly upstreamStatus: number | null,
-  ) {
-    super(message);
-    this.name = "ProviderUpstreamInvokeError";
-  }
-}
-
-export class InvocationInvalidArgsError extends Error {
-  readonly kind = "invalid_args" as const;
-  constructor(message: string) {
-    super(message);
-    this.name = "InvocationInvalidArgsError";
-  }
-}
-
-interface ErrorEnvelope {
-  detail?: {
-    error?: string;
-    message?: string;
-    posture?: string;
-    required_role?: string;
-    actor_role?: string;
-    reason?: string;
-    plugin?: string;
-    skill?: string;
-    capability?: string;
-    blocked_reason?: string;
-    gate_state?: Record<string, unknown>;
-    provider?: string;
-    code?: string;
-    upstream_status?: number;
-  };
-}
-
-async function readInvocationEnvelope(
-  res: Response,
-): Promise<ErrorEnvelope> {
-  try {
-    return (await res.json()) as ErrorEnvelope;
-  } catch {
-    return {};
-  }
-}
-
-export const invokeCapability = async (
-  slug: string,
-  body: { module_id: string; capability_id: string; args?: Record<string, unknown> },
-): Promise<InvocationResponse> => {
-  const res = await apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/invocations`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ args: {}, ...body }),
-    },
-  );
-  if (res.ok) return jsonOrThrow<InvocationResponse>(res);
-
-  const env = await readInvocationEnvelope(res);
-  const d = env.detail ?? {};
-  switch (d.error) {
-    case "posture_gate_blocked":
-      throw new PostureBlockedError(
-        d.message ?? "Posture gate blocked invocation.",
-        d.posture ?? "unknown",
-        d.required_role ?? "unknown",
-        d.actor_role ?? "unknown",
-        d.reason ?? "posture_gate_failed",
-      );
-    case "capability_denied":
-      throw new CapabilityDeniedError(
-        d.message ?? "Capability denied.",
-        d.plugin ?? "",
-        d.skill ?? "",
-        d.capability ?? "",
-      );
-    case "phase1_blocked":
-      throw new Phase1BlockedError(
-        d.message ?? "Advice-boundary gate blocked invocation.",
-        d.blocked_reason ?? "unknown",
-        d.gate_state ?? {},
-      );
-    case "provider_key_missing":
-      throw new ProviderKeyMissingForInvokeError(
-        d.message ?? "Provider API key not configured.",
-        d.provider ?? null,
-      );
-    case "provider_upstream_error":
-      throw new ProviderUpstreamInvokeError(
-        d.message ?? "Provider upstream error.",
-        d.provider ?? null,
-        d.code ?? null,
-        d.upstream_status ?? null,
-      );
-    case "invalid_args":
-      throw new InvocationInvalidArgsError(
-        d.message ?? "Capability rejected the args.",
-      );
-    default:
-      // Unknown structured error — fall back to a plain Error with the
-      // full envelope so the runner UI can surface raw substrate text.
-      throw new Error(
-        `Invocation failed (${res.status}): ${d.error ?? "unknown"} — ${d.message ?? ""}`,
-      );
-  }
-};
-
-// Matter artifacts
-// ---------------------------------------------------------------------------
-// Guided Demo Loop v1 — keyless end-to-end proof
-// ---------------------------------------------------------------------------
-
-export interface GuidedDemoHandles {
-  matter_slug: string;
-  matter_title: string;
-  module_id: string;
-  capability_id: string;
-  document_id: string;
-  document_filename: string;
-  model_id: string;
-}
-
-export const ensureGuidedLoop = () =>
-  apiFetch(`${API}/demo/guided-loop`, { method: "POST" }).then((r) =>
-    jsonOrThrow<GuidedDemoHandles>(r),
-  );
-
-export interface ArtifactSummary {
-  id: string;
-  matter_id: string;
-  module_id: string;
-  capability_id: string;
-  invocation_id: string;
-  kind: string;
-  created_by_id: string;
-  created_at: string;
-  size_bytes: number;
-}
-
-export interface ArtifactRead extends ArtifactSummary {
-  payload: Record<string, unknown>;
-}
-
-export const listArtifacts = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/artifacts`).then((r) =>
-    jsonOrThrow<ArtifactSummary[]>(r),
-  );
-
-export const readArtifact = (slug: string, artifactId: string) =>
-  apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/artifacts/${encodeURIComponent(artifactId)}`,
-  ).then((r) => jsonOrThrow<ArtifactRead>(r));
-
-// ---------------------------------------------------------------------------
-// Supervisor Review v1 — review/approval over a matter artifact
-// ---------------------------------------------------------------------------
-
-export type ReviewState =
-  | "pending"
-  | "approved"
-  | "rejected"
-  | "changes_requested"
-  | "overridden";
-
-export type ReviewDecision = "approve" | "reject" | "request_changes" | "override";
-
-export interface SupervisorReview {
-  id: string;
-  matter_id: string;
-  artifact_id: string;
-  invocation_id: string;
-  module_id: string;
-  capability_id: string;
-  kind: string;
-  artifact_hash: string;
-  state: ReviewState;
-  requested_by_id: string;
-  requested_at: string;
-  decided_by_id: string | null;
-  decided_at: string | null;
-  note: string | null;
-}
-
-export const listSupervisorReviews = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/reviews`).then((r) =>
-    jsonOrThrow<{ matter_id: string; reviews: SupervisorReview[] }>(r),
-  );
-
-// Artifact kinds the backend supervisor-review API accepts. Mirrors
-// `backend/app/models/matter_review.py:REVIEW_ELIGIBLE_KINDS`. Keep these
-// in sync — the frontend only offers "Request review" for kinds the
-// backend will accept (no dead buttons).
-export const REVIEW_ELIGIBLE_KINDS: readonly string[] = [
-  "findings_pack",
-  "skill_response",
-];
-
-export const requestReview = (slug: string, artifactId: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/reviews`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ artifact_id: artifactId }),
-  }).then((r) => jsonOrThrow<SupervisorReview>(r));
-
-// ---------------------------------------------------------------------------
-// Professional Sign-Off v1 — author sign-off over a matter artifact
-// ---------------------------------------------------------------------------
-
-export type SignoffDecision = "signed" | "signed_with_observations" | "rejected";
-
-export interface Signoff {
-  id: string;
-  matter_id: string;
-  artifact_id: string;
-  invocation_id: string;
-  module_id: string;
-  capability_id: string;
-  kind: string;
-  artifact_hash: string;
-  decision: SignoffDecision;
-  reasoning: string | null;
-  signer_id: string;
-  signer_email: string | null;
-  signer_is_author: boolean;
-  signed_at: string;
-  is_current: boolean;
-}
-
-export const createSignoff = (
-  slug: string,
-  body: { artifact_id: string; decision: SignoffDecision; reasoning?: string },
-) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/signoffs`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => jsonOrThrow<Signoff>(r));
-
-export const listSignoffs = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/signoffs`).then((r) =>
-    jsonOrThrow<{ matter_id: string; signoffs: Signoff[] }>(r),
-  );
-
-export const getSignoff = (slug: string, signoffId: string) =>
-  apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/signoffs/${encodeURIComponent(signoffId)}`,
-  ).then((r) => jsonOrThrow<Signoff>(r));
-
-export const decideReview = (
-  slug: string,
-  reviewId: string,
-  decision: ReviewDecision,
-  note?: string,
-) =>
-  apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/reviews/${encodeURIComponent(reviewId)}/decide`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ decision, note: note ?? null }),
-    },
-  ).then((r) => jsonOrThrow<SupervisorReview>(r));
-
-// ---------------------------------------------------------------------------
-// Reconstruction
-// ---------------------------------------------------------------------------
-
-// The three legal source values per backend/app/core/audit_reconstruction.py.
-export type ReconstructionSource = "audit" | "state_machine" | "advice_boundary";
-
-export const ALL_RECONSTRUCTION_SOURCES: ReconstructionSource[] = [
-  "audit",
-  "state_machine",
-  "advice_boundary",
-];
-
-export interface TimelineActor {
-  user_id?: string;
-  role?: string;
-  email?: string;
-  // Substrate-side actor dict can carry arbitrary keys.
-  [k: string]: unknown;
-}
-
-export interface TimelineEntry {
-  source: ReconstructionSource;
-  occurred_at: string;
-  action: string;
-  actor: TimelineActor;
-  matter_id: string | null;
-  module_id: string | null;
-  capability_id: string | null;
-  payload: Record<string, unknown>;
-  refs: Record<string, unknown>;
-  source_row_id: string;
-}
-
-export interface ReconstructionResponse {
-  entries: TimelineEntry[];
-  next_cursor: string | null;
-  total_in_window_estimate: number;
-}
-
-export interface ReconstructionOptions {
-  since?: string;
-  until?: string;
-  include?: ReconstructionSource[];
-  cursor?: string;
-  limit?: number;
-  // Substrate-side filters. Earlier the frontend filtered these
-  // client-side, which produced false-negatives on dense matter
-  // timelines. They're now server-pushdown filters that apply
-  // BEFORE pagination.
-  invocation_id?: string;
-  action?: string;
-}
-
-// ---------------------------------------------------------------------------
-// Admin users
-// ---------------------------------------------------------------------------
-
-// Locked vocabulary — substrate ALLOWED_ROLES at admin_users.py:52.
-// workspace_admin is a settable role; it does NOT bypass posture gates
-// (only qualified_solicitor satisfies B_mixed — substrate truth from
-// posture_gate.py POSTURE_POLICY; the two axes are independent).
-export const ALLOWED_ROLES = [
-  "solicitor",
-  "qualified_solicitor",
-  "workspace_admin",
-] as const;
-export type UserRole = (typeof ALLOWED_ROLES)[number];
-
-export interface UserAdminRead {
-  id: string;
-  email: string;
-  role: string;
-  is_superuser: boolean;
-  is_active: boolean;
-  is_verified: boolean;
-  name: string;
-  created_at: string | null;
-}
-
-export interface UserRoleOut {
-  id: string;
-  email: string;
-  role: string;
-  is_superuser: boolean;
-}
-
-export interface ListAdminUsersOptions {
-  role?: UserRole;
-  is_superuser?: boolean;
-}
-
-export class AdminRequiredError extends Error {
-  readonly kind = "admin_required" as const;
-  constructor(message: string) {
-    super(message);
-    this.name = "AdminRequiredError";
-  }
-}
-
-export class SelfPromotionForbiddenError extends Error {
-  readonly kind = "self_promotion_forbidden" as const;
-  constructor(message: string) {
-    super(message);
-    this.name = "SelfPromotionForbiddenError";
-  }
-}
-
-export class InvalidRoleError extends Error {
-  readonly kind = "invalid_role" as const;
-  constructor(
-    message: string,
-    public readonly supplied: string,
-    public readonly allowed: string[],
-  ) {
-    super(message);
-    this.name = "InvalidRoleError";
-  }
-}
-
-async function readEnv(res: Response): Promise<{
-  detail?: {
-    error?: string;
-    message?: string;
-    supplied?: string;
-    allowed?: string[];
-  };
-}> {
-  try {
-    return (await res.json()) as {
-      detail?: {
-        error?: string;
-        message?: string;
-        supplied?: string;
-        allowed?: string[];
-      };
-    };
-  } catch {
-    return {};
-  }
-}
-
-export const listAdminUsers = async (
-  opts: ListAdminUsersOptions = {},
-): Promise<UserAdminRead[]> => {
-  const params = new URLSearchParams();
-  if (opts.role !== undefined) params.set("role", opts.role);
-  if (opts.is_superuser !== undefined) {
-    params.set("is_superuser", String(opts.is_superuser));
-  }
-  const qs = params.toString();
-  const res = await apiFetch(`${API}/admin/users${qs ? `?${qs}` : ""}`);
-  if (res.status === 403) {
-    const env = await readEnv(res);
-    throw new AdminRequiredError(
-      env.detail?.message ?? "Endpoint requires superuser.",
-    );
-  }
-  return jsonOrThrow<UserAdminRead[]>(res);
-};
-
-export const getAdminUser = async (userId: string): Promise<UserAdminRead> => {
-  const res = await apiFetch(`${API}/admin/users/${encodeURIComponent(userId)}`);
-  if (res.status === 403) {
-    const env = await readEnv(res);
-    throw new AdminRequiredError(
-      env.detail?.message ?? "Endpoint requires superuser.",
-    );
-  }
-  return jsonOrThrow<UserAdminRead>(res);
-};
-
-// POST body is {role} ONLY — substrate RoleChangeRequest at
-// admin_users.py:57. Operator-supplied "reason" is a backend
-// concern, not a frontend invention.
-export const changeUserRole = async (
-  userId: string,
-  role: UserRole,
-): Promise<UserRoleOut> => {
-  const res = await apiFetch(
-    `${API}/admin/users/${encodeURIComponent(userId)}/role`,
-    {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ role }),
-    },
-  );
-  if (res.status === 403) {
-    const env = await readEnv(res);
-    if (env.detail?.error === "self_promotion_forbidden") {
-      throw new SelfPromotionForbiddenError(
-        env.detail?.message ??
-          "Superusers cannot change their own role via this endpoint.",
-      );
-    }
-    throw new AdminRequiredError(
-      env.detail?.message ?? "Endpoint requires superuser.",
-    );
-  }
-  if (res.status === 422) {
-    const env = await readEnv(res);
-    if (env.detail?.error === "invalid_role") {
-      throw new InvalidRoleError(
-        `Role ${env.detail?.supplied ?? "?"} not in allowed set.`,
-        env.detail?.supplied ?? role,
-        env.detail?.allowed ?? Array.from(ALLOWED_ROLES),
-      );
-    }
-  }
-  return jsonOrThrow<UserRoleOut>(res);
-};
-
-// Workspace / admin reconstruction. Same shape as
-// the matter endpoint; no slug. Substrate gates on superuser; UI
-// also gates upstream to avoid pointless 403s.
-export const getAdminReconstruction = (
-  opts: ReconstructionOptions = {},
-): Promise<ReconstructionResponse> => {
-  const params = new URLSearchParams();
-  if (opts.since) params.set("since", opts.since);
-  if (opts.until) params.set("until", opts.until);
-  if (opts.include && opts.include.length > 0) {
-    params.set("include", opts.include.join(","));
-  }
-  if (opts.cursor) params.set("cursor", opts.cursor);
-  if (opts.limit !== undefined) params.set("limit", String(opts.limit));
-  if (opts.invocation_id) params.set("invocation_id", opts.invocation_id);
-  if (opts.action) params.set("action", opts.action);
-  const qs = params.toString();
-  return apiFetch(
-    `${API}/admin/audit/reconstruction${qs ? `?${qs}` : ""}`,
-  ).then((r) => jsonOrThrow<ReconstructionResponse>(r));
-};
-
-export const getReconstruction = (
-  slug: string,
-  opts: ReconstructionOptions = {},
-): Promise<ReconstructionResponse> => {
-  const params = new URLSearchParams();
-  if (opts.since) params.set("since", opts.since);
-  if (opts.until) params.set("until", opts.until);
-  if (opts.include && opts.include.length > 0) {
-    params.set("include", opts.include.join(","));
-  }
-  if (opts.cursor) params.set("cursor", opts.cursor);
-  if (opts.limit !== undefined) params.set("limit", String(opts.limit));
-  if (opts.invocation_id) params.set("invocation_id", opts.invocation_id);
-  if (opts.action) params.set("action", opts.action);
-  const qs = params.toString();
-  return apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/audit/reconstruction${qs ? `?${qs}` : ""}`,
-  ).then((r) => jsonOrThrow<ReconstructionResponse>(r));
-};
-
-export const listMatters = () =>
-  apiFetch(`${API}/matters`).then((r) => jsonOrThrow<Matter[]>(r));
-
-export const getMatter = (slug: string) =>
-  apiFetch(`${API}/matters/${slug}`).then((r) => jsonOrThrow<Matter>(r));
-
-export const createMatter = (body: MatterCreate) =>
-  apiFetch(`${API}/matters`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => jsonOrThrow<Matter>(r));
-
-export const listDocuments = (slug: string) =>
-  apiFetch(`${API}/matters/${slug}/documents`).then((r) => jsonOrThrow<MatterDocument[]>(r));
-
-// Typed upload error. Lets the UI show a friendly inline banner for
-// the three validation failures the backend enforces: unsupported MIME
-// (415), declared MIME doesn't match magic bytes (415), and over the
-// 25 MB cap (413). Anything else flows through as a generic Error
-// from `jsonOrThrow`.
-export type UploadErrorKind =
-  | "unsupported_mime"
-  | "magic_byte_mismatch"
-  | "upload_too_large";
-
-export class UploadError extends Error {
-  kind: UploadErrorKind;
-  status: number;
-  constructor(kind: UploadErrorKind, status: number, message: string) {
-    super(message);
-    this.name = "UploadError";
-    this.kind = kind;
-    this.status = status;
-  }
-}
-
-function formatMb(bytes: number): string {
-  return `${(bytes / 1024 / 1024).toFixed(1)} MB`;
-}
-
-export const uploadDocument = async (
-  slug: string,
-  file: File,
-  tag?: string,
-  fromDisclosure?: boolean,
-): Promise<MatterDocument> => {
-  const fd = new FormData();
-  fd.append("file", file);
-  if (tag) fd.append("tag", tag);
-  if (fromDisclosure) fd.append("from_disclosure", "true");
-  const res = await apiFetch(`${API}/matters/${slug}/documents`, {
-    method: "POST",
-    body: fd,
-  });
-  if (res.status === 413 || res.status === 415) {
-    let detail: Record<string, unknown> | null = null;
-    try {
-      const body = (await res.json()) as { detail?: Record<string, unknown> };
-      detail = body?.detail ?? null;
-    } catch {
-      detail = null;
-    }
-    if (res.status === 415) {
-      const errKind = (detail?.error as string | undefined) ?? "unsupported_mime";
-      if (errKind === "magic_byte_mismatch") {
-        const declared = (detail?.declared_mime as string | null) || file.type || "the declared type";
-        const inferred = (detail?.inferred_format as string | null) ?? "something else";
-        throw new UploadError(
-          "magic_byte_mismatch",
-          415,
-          `File contents do not match its declared type. Declared as ${declared}; the bytes look like ${inferred}. Re-export from the source app and try again.`,
-        );
-      }
-      const got = (detail?.got as string | null) || file.type || "unknown";
-      throw new UploadError(
-        "unsupported_mime",
-        415,
-        `That file type is not supported (${got}). Upload a PDF, DOCX, DOC, TXT, MD, or RTF.`,
-      );
-    }
-    const maxBytes = Number(detail?.max_bytes ?? 25 * 1024 * 1024);
-    const gotBytes = Number(detail?.got_bytes ?? file.size);
-    throw new UploadError(
-      "upload_too_large",
-      413,
-      `File is too large (${formatMb(gotBytes)}). The limit is ${formatMb(maxBytes)} per document.`,
-    );
-  }
-  return jsonOrThrow<MatterDocument>(res);
-};
-
-export interface AuditEntry {
-  id: string;
-  timestamp: string;
-  actor_id: string | null;
-  matter_id: string | null;
-  action: string;
-  module: string | null;
-  resource_type: string | null;
-  resource_id: string | null;
-  model_used: string | null;
-  prompt_hash: string | null;
-  response_hash: string | null;
-  token_count: number | null;
-  latency_ms: number | null;
-  payload: Record<string, unknown>;
-}
-
-export const listAudit = (slug: string, limit = 50) =>
-  apiFetch(`${API}/matters/${slug}/audit?limit=${limit}`).then((r) => jsonOrThrow<AuditEntry[]>(r));
-
-// Audit hash-chain verification (notary-minimal). Substrate truth:
-// GET /api/matters/{slug}/audit/chain recomputes every link in the
-// matter scope; `head.chain_hash` is the record's fingerprint.
-export interface AuditChainHead {
-  chain_hash: string;
-  scope_sequence: number;
-  entry_hash: string;
-}
-
-export interface AuditChainIssue {
-  code: string;
-  message: string;
-  audit_entry_id: string | null;
-  chain_id: number | null;
-}
-
-export interface AuditChainStatus {
-  verified: boolean;
-  scope: string;
-  length: number;
-  head: AuditChainHead | null;
-  issues: AuditChainIssue[];
-}
-
-export const getAuditChainStatus = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/audit/chain`).then((r) =>
-    jsonOrThrow<AuditChainStatus>(r),
-  );
-
-export const setPrivilege = (slug: string, posture: string) =>
-  apiFetch(`${API}/matters/${slug}/privilege`, {
-    method: "PATCH",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ privilege_posture: posture }),
-  }).then((r) => jsonOrThrow<Matter>(r));
-
-export class AccountHasMattersError extends Error {
-  readonly matterCount: number;
-  constructor(matterCount: number) {
-    super(
-      `Account owns ${matterCount} matter${matterCount === 1 ? "" : "s"}. ` +
-        "Export or delete matters before deleting the account.",
-    );
-    this.name = "AccountHasMattersError";
-    this.matterCount = matterCount;
-  }
-}
-
-export const deleteAccount = async (): Promise<void> => {
-  const r = await apiFetch(`${AUTH}/users/me`, { method: "DELETE" });
-  if (r.status === 204) return;
-  if (r.status === 409) {
-    let count = 0;
-    try {
-      const body = (await r.json()) as { detail?: { matter_count?: number } };
-      count = body.detail?.matter_count ?? 0;
-    } catch {
-      // body parse failure — leave count at 0 and surface the generic message.
-    }
-    throw new AccountHasMattersError(count);
-  }
-  throw new Error(`deleteAccount: ${r.status} ${r.statusText}`);
-};
-
-export interface ChronologyEvent {
-  id: string;
-  event_date: string;
-  description: string;
-  significance: number;
-  source_doc_ids: string[];
-  source_doc_filenames: string[];
-  priv_flag: boolean;
-  from_disclosure: boolean;
-  proceedings_refs: string[];
-  created_at: string;
-  redacted: boolean;
-}
-
-export interface GateState {
-  required: boolean;
-  confirmed: boolean;
-  confirmed_at: string | null;
-  tainted_event_count: number;
-}
-
-export interface ChronologyResponse {
-  matter_slug: string;
-  events: ChronologyEvent[];
-  gate: GateState;
-  statement_of_facts_variant: ChronologyEvent[];
-}
-
-export const getChronology = (slug: string) =>
-  apiFetch(`${API}/matters/${slug}/chronology`).then((r) => jsonOrThrow<ChronologyResponse>(r));
-
-export const confirmGate = (slug: string, acknowledgement: string) =>
-  apiFetch(`${API}/matters/${slug}/chronology/gate`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ acknowledgement }),
-  }).then((r) => jsonOrThrow<GateState>(r));
-
-// ----- Document body + edit instructions ---------------------------------
-
-export interface DocumentBody {
-  document_id: string;
-  kind: string;
-  extracted_text: string;
-  extraction_method: string;
-  extracted_at: string;
-  char_count: number;
-  page_count: number | null;
-  error_reason: string | null;
-}
-
-export interface DocumentWorkspaceBlock {
-  id: string;
-  type: "paragraph" | "table_cell";
-  ordinal: number;
-  text: string;
-}
-
-export interface DocumentWorkspace {
-  document_id: string;
-  filename: string;
-  mime_type: string;
-  source: "original_docx" | "latest_version" | "extracted_body";
-  source_version_id: string | null;
-  source_version_number: number | null;
-  extraction_method: string | null;
-  blocks: DocumentWorkspaceBlock[];
-  text: string;
-  char_count: number;
-  notes: string[];
-}
-
-export type EditMode =
-  | "tighten"
-  | "rewrite"
-  | "summarise"
-  | "free-text"
-  | "uk-jurisdiction-sweep";
-
-export interface DocumentVersionRead {
-  id: string;
-  document_id: string;
-  version_number: number;
-  kind: string;
-  created_by_id: string;
-  created_at: string;
-  storage_uri: string | null;
-  filename: string | null;
-  mime_type: string | null;
-  size_bytes: number | null;
-  sha256: string | null;
-  notes: string | null;
-  resolved_text: string | null;
-  resolved_json?: Record<string, unknown> | null;
-}
-
-export interface DocumentEditRead {
-  id: string;
-  document_version_id: string;
-  change_id: string;
-  correlation_id: string | null;
-  deleted_text: string;
-  inserted_text: string;
-  context_before: string;
-  context_after: string;
-  rationale: string | null;
-  status: string;
-  created_at: string;
-}
-
-export interface EditInstructionResponse {
-  version: DocumentVersionRead;
-  pending_edits: DocumentEditRead[];
-  model_used: string;
-  model_notes: string;
-  instruction_hash: string;
-  parse_ok: boolean;
-}
-
-export const getDocumentBody = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/body`).then((r) =>
-    jsonOrThrow<DocumentBody>(r),
-  );
-
-export const getDocumentWorkspace = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/workspace`).then((r) =>
-    jsonOrThrow<DocumentWorkspace>(r),
-  );
-
-export const saveDocumentWorkspace = (
-  documentId: string,
-  text: string,
-  notes?: string,
-  resolvedJson?: Record<string, unknown> | null,
-) =>
-  apiFetch(`${API}/documents/${documentId}/workspace`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      text,
-      notes: notes?.trim() || null,
-      resolved_json: resolvedJson ?? null,
-    }),
-  }).then((r) =>
-    resolutionJsonOrThrow<{
-      version: DocumentVersionRead;
-      workspace: DocumentWorkspace;
-    }>(r),
-  );
-
-// Original File Retrieval v1 — browser-navigable URL for the streamed
-// backend proxy. Used directly as an <a href> (open inline) or with
-// download=1 (attachment); the browser handles the response, so this
-// returns a URL rather than fetching bytes through React state.
-export const documentOriginalUrl = (
-  documentId: string,
-  opts?: { download?: boolean },
-): string =>
-  `${API}/documents/${encodeURIComponent(documentId)}/original${
-    opts?.download ? "?download=1" : ""
-  }`;
-
-export async function fetchDocumentOriginalBlob(documentId: string): Promise<Blob> {
-  const resp = await apiFetch(documentOriginalUrl(documentId));
-  if (!resp.ok) {
-    const text = await resp.text().catch(() => "");
-    throw new Error(text || `Could not load original document (${resp.status})`);
-  }
-  return resp.blob();
-}
-
-export const documentVersionDocxUrl = (
-  documentId: string,
-  versionId: string,
-): string =>
-  `${API}/documents/${encodeURIComponent(documentId)}/versions/${encodeURIComponent(
-    versionId,
-  )}/docx`;
-
-export const documentVersionPdfUrl = (
-  documentId: string,
-  versionId: string,
-): string =>
-  `${API}/documents/${encodeURIComponent(documentId)}/versions/${encodeURIComponent(
-    versionId,
-  )}/pdf`;
-
-export const documentVersionOriginalUrl = (
-  documentId: string,
-  versionId: string,
-  opts?: { download?: boolean },
-): string =>
-  `${API}/documents/${encodeURIComponent(documentId)}/versions/${encodeURIComponent(
-    versionId,
-  )}/original${opts?.download ? "?download=1" : ""}`;
-
-// ---------------------------------------------------------------------------
-// Matter lifecycle + export (LMF UX v1) — over the stable LMF endpoints
-// ---------------------------------------------------------------------------
-
-export type JobStatus = "queued" | "running" | "succeeded" | "failed" | "cancelled";
-
-export interface JobRead {
-  id: string;
-  matter_id: string;
-  kind: string;
-  status: JobStatus;
-  stage: string | null;
-  progress: number | null;
-  error_code: string | null;
-  error_message: string | null;
-  created_at: string | null;
-  started_at: string | null;
-  finished_at: string | null;
-  result_payload: Record<string, unknown> | null;
-}
-
-export const createMatterExport = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/export`, {
-    method: "POST",
-  }).then((r) => jsonOrThrow<JobRead>(r));
-
-export const getJob = (slug: string, jobId: string) =>
-  apiFetch(
-    `${API}/matters/${encodeURIComponent(slug)}/jobs/${encodeURIComponent(jobId)}`,
-  ).then((r) => jsonOrThrow<JobRead>(r));
-
-// Browser-navigable download URL for a completed export (302 presigned
-// on S3, or a streamed attachment locally — the browser handles it).
-export const matterExportDownloadUrl = (slug: string, jobId: string): string =>
-  `${API}/matters/${encodeURIComponent(slug)}/export/${encodeURIComponent(jobId)}`;
-
-export const closeMatter = (slug: string) =>
-  apiFetch(`${API}/matters/${encodeURIComponent(slug)}/close`, {
-    method: "POST",
-  }).then((r) => jsonOrThrow<Matter>(r));
-
-export const deleteMatter = async (slug: string): Promise<void> => {
-  const res = await apiFetch(`${API}/matters/${encodeURIComponent(slug)}`, {
-    method: "DELETE",
-  });
-  if (!res.ok && res.status !== 204) {
-    const text = await res.text();
-    throw new Error(`${res.status} ${res.statusText}: ${text}`);
-  }
-};
-
-export const postEditInstruction = (
-  documentId: string,
-  instruction: string,
-  mode: EditMode,
-) =>
-  apiFetch(`${API}/documents/${documentId}/edit-instructions`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ instruction, mode }),
-  }).then((r) => jsonOrThrow<EditInstructionResponse>(r));
-
-// ----- Generated .docx export --------------------------------------------
-
-export interface GeneratedDocxResponse {
-  file_uuid: string;
-  storage_uri: string;
-  byte_count: number;
-  download_url: string;
-}
-
-export async function downloadGeneratedDocx(fileUuid: string): Promise<Blob> {
-  const resp = await apiFetch(`${API}/documents/generated/${fileUuid}`);
-  if (!resp.ok) {
-    const text = await resp.text();
-    throw new Error(`${resp.status} ${resp.statusText}: ${text}`);
-  }
-  return resp.blob();
-}
-
-// ----- Tracked changes accept/reject -------------------------------------
-
-export interface EditResolutionResponse {
-  edit: DocumentEditRead;
-  new_version: DocumentVersionRead | null;
-  resolved_text: string | null;
-}
-
-export interface BulkResolutionResponse {
-  affected_count: number;
-  new_version: DocumentVersionRead;
-  resolved_text: string;
-}
-
-export interface DocumentVersionSummary {
-  version: DocumentVersionRead;
-  pending_count: number;
-  accepted_count: number;
-  rejected_count: number;
-}
-
-export interface DocumentCommentRead {
-  id: string;
-  document_id: string;
-  author_id: string;
-  quote_text: string | null;
-  body_sha256: string | null;
-  anchor_start: number | null;
-  anchor_end: number | null;
-  body: string;
-  status: "open" | "resolved";
-  created_at: string;
-  resolved_at: string | null;
-  resolved_by_id: string | null;
-}
-
-export interface DocumentEditSessionRead {
-  id: string;
-  document_id: string;
-  user_id: string;
-  client_id: string;
-  user_label: string;
-  started_at: string;
-  last_seen_at: string;
-  ended_at: string | null;
-}
-
-export interface DocumentEditSessionResponse {
-  current: DocumentEditSessionRead;
-  active: DocumentEditSessionRead[];
-}
-
-export interface DocumentWorkingDraftRead {
-  document_id: string;
-  updated_by_id: string | null;
-  updated_at: string | null;
-  plain_text: string;
-  editor_json: Record<string, unknown> | null;
-  base_version_id: string | null;
-  version_counter: number;
-  client_id: string | null;
-}
-
-export interface DocumentAssetUploadRead {
-  id: string;
-  filename: string;
-  mime_type: string;
-  size_bytes: number;
-  sha256: string;
-  url: string;
-}
-
-export class ConflictError extends Error {
-  status = 409;
-  detail: unknown;
-  constructor(message = "edit already resolved", detail: unknown = null) {
-    super(message);
-    this.name = "ConflictError";
-    this.detail = detail;
-  }
-}
-
-async function resolutionJsonOrThrow<T>(res: Response): Promise<T> {
-  if (res.status === 409) {
-    const text = await res.text();
-    try {
-      const parsed = JSON.parse(text) as unknown;
-      const detail =
-        parsed && typeof parsed === "object" && "detail" in parsed
-          ? (parsed as { detail?: unknown }).detail
-          : parsed;
-      const message =
-        detail && typeof detail === "object" && "message" in detail
-          ? String((detail as { message?: unknown }).message || "")
-          : "";
-      throw new ConflictError(message || "edit already resolved", detail);
-    } catch (err) {
-      if (err instanceof ConflictError) throw err;
-      throw new ConflictError(text || "edit already resolved");
-    }
-  }
-  if (!res.ok) {
-    const text = await res.text();
-    throw new Error(`${res.status} ${res.statusText}: ${text}`);
-  }
-  return res.json() as Promise<T>;
-}
-
-export interface PendingEditsResponse {
-  version: EditInstructionResponse["version"] | null;
-  pending_edits: EditInstructionResponse["pending_edits"];
-}
-
-// Pending redlines survive reload: the substrate stores proposed edits
-// in document_edits; this rehydrates them on document open.
-export const getPendingEdits = (documentId: string) =>
-  apiFetch(`${API}/documents/${encodeURIComponent(documentId)}/edits/pending`).then(
-    (r) => jsonOrThrow<PendingEditsResponse>(r),
-  );
-
-export const acceptEdit = (editId: string) =>
-  apiFetch(`${API}/documents/edits/${editId}/accept`, { method: "POST" }).then(
-    (r) => resolutionJsonOrThrow<EditResolutionResponse>(r),
-  );
-
-export const rejectEdit = (editId: string) =>
-  apiFetch(`${API}/documents/edits/${editId}/reject`, { method: "POST" }).then(
-    (r) => resolutionJsonOrThrow<EditResolutionResponse>(r),
-  );
-
-export const acceptAll = (versionId: string) =>
-  apiFetch(`${API}/documents/versions/${versionId}/accept-all`, {
-    method: "POST",
-  }).then((r) => resolutionJsonOrThrow<BulkResolutionResponse>(r));
-
-export const rejectAll = (versionId: string) =>
-  apiFetch(`${API}/documents/versions/${versionId}/reject-all`, {
-    method: "POST",
-  }).then((r) => resolutionJsonOrThrow<BulkResolutionResponse>(r));
-
-export const getDocumentVersions = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/versions`).then((r) =>
-    resolutionJsonOrThrow<DocumentVersionSummary[]>(r),
-  );
-
-export const getDocumentWorkingDraft = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/draft`).then((r) =>
-    resolutionJsonOrThrow<DocumentWorkingDraftRead>(r),
-  );
-
-export const saveDocumentWorkingDraft = (
-  documentId: string,
-  body: {
-    plain_text: string;
-    editor_json?: Record<string, unknown> | null;
-    base_version_id?: string | null;
-    client_id?: string | null;
-    expected_version_counter?: number | null;
-  },
-) =>
-  apiFetch(`${API}/documents/${documentId}/draft`, {
-    method: "PUT",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => resolutionJsonOrThrow<DocumentWorkingDraftRead>(r));
-
-export const commitDocumentWorkingDraft = (
-  documentId: string,
-  notes?: string,
-  clearDraft = true,
-  expectedVersionCounter?: number | null,
-) =>
-  apiFetch(`${API}/documents/${documentId}/draft/commit`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      notes: notes?.trim() || null,
-      clear_draft: clearDraft,
-      expected_version_counter: expectedVersionCounter ?? null,
-    }),
-  }).then((r) => resolutionJsonOrThrow<DocumentVersionRead>(r));
-
-export const saveDocumentVersion = (
-  documentId: string,
-  resolvedText: string,
-  notes?: string,
-  resolvedJson?: Record<string, unknown> | null,
-) =>
-  apiFetch(`${API}/documents/${documentId}/versions/manual`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      resolved_text: resolvedText,
-      resolved_json: resolvedJson ?? null,
-      notes,
-    }),
-  }).then((r) => resolutionJsonOrThrow<DocumentVersionRead>(r));
-
-export const uploadDocumentVersion = async (
-  documentId: string,
-  file: File,
-  notes?: string,
-): Promise<DocumentVersionRead> => {
-  const fd = new FormData();
-  fd.append("file", file);
-  if (notes?.trim()) fd.append("notes", notes.trim());
-  const res = await apiFetch(`${API}/documents/${documentId}/versions/upload`, {
-    method: "POST",
-    body: fd,
-  });
-  if (res.status === 413 || res.status === 415) {
-    let detail: Record<string, unknown> | null = null;
-    try {
-      const body = (await res.json()) as { detail?: Record<string, unknown> };
-      detail = body?.detail ?? null;
-    } catch {
-      detail = null;
-    }
-    if (res.status === 415) {
-      const errKind = (detail?.error as string | undefined) ?? "unsupported_mime";
-      if (errKind === "magic_byte_mismatch") {
-        const declared = (detail?.declared_mime as string | null) || file.type || "the declared type";
-        const inferred = (detail?.inferred_format as string | null) ?? "something else";
-        throw new UploadError(
-          "magic_byte_mismatch",
-          415,
-          `File contents do not match its declared type. Declared as ${declared}; the bytes look like ${inferred}. Re-export from the source app and try again.`,
-        );
-      }
-      const got = (detail?.got as string | null) || file.type || "unknown";
-      throw new UploadError(
-        "unsupported_mime",
-        415,
-        `That file type is not supported (${got}). Upload a PDF, DOCX, DOC, TXT, MD, or RTF.`,
-      );
-    }
-    const maxBytes = Number(detail?.max_bytes ?? 25 * 1024 * 1024);
-    const gotBytes = Number(detail?.got_bytes ?? file.size);
-    throw new UploadError(
-      "upload_too_large",
-      413,
-      `File is too large (${formatMb(gotBytes)}). The limit is ${formatMb(maxBytes)} per document.`,
-    );
-  }
-  return resolutionJsonOrThrow<DocumentVersionRead>(res);
-};
-
-export const uploadDocumentAsset = async (
-  documentId: string,
-  file: File,
-): Promise<DocumentAssetUploadRead> => {
-  const fd = new FormData();
-  fd.append("file", file);
-  const res = await apiFetch(`${API}/documents/${documentId}/assets`, {
-    method: "POST",
-    body: fd,
-  });
-  if (res.status === 413 || res.status === 415) {
-    if (res.status === 415) {
-      throw new UploadError(
-        "unsupported_mime",
-        415,
-        "That image type is not supported. Upload PNG, JPEG, WebP, or GIF.",
-      );
-    }
-    throw new UploadError(
-      "upload_too_large",
-      413,
-      `Image is too large (${formatMb(file.size)}). The limit is ${formatMb(
-        5 * 1024 * 1024,
-      )} per image.`,
-    );
-  }
-  const uploaded = await resolutionJsonOrThrow<DocumentAssetUploadRead>(res);
-  return {
-    ...uploaded,
-    url: uploaded.url.startsWith("/api/") ? `${BACKEND_ROOT}${uploaded.url}` : uploaded.url,
-  };
-};
-
-export const restoreDocumentVersion = (
-  documentId: string,
-  versionId: string,
-  notes?: string,
-) =>
-  apiFetch(`${API}/documents/${documentId}/versions/${versionId}/restore`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ notes: notes?.trim() || null }),
-  }).then((r) => resolutionJsonOrThrow<DocumentVersionRead>(r));
-
-export const getDocumentComments = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/comments`).then((r) =>
-    resolutionJsonOrThrow<DocumentCommentRead[]>(r),
-  );
-
-export const getDocumentEditSessions = (documentId: string) =>
-  apiFetch(`${API}/documents/${documentId}/edit-sessions`).then((r) =>
-    resolutionJsonOrThrow<DocumentEditSessionRead[]>(r),
-  );
-
-export const startDocumentEditSession = (
-  documentId: string,
-  clientId: string,
-) =>
-  apiFetch(`${API}/documents/${documentId}/edit-sessions`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ client_id: clientId }),
-  }).then((r) => resolutionJsonOrThrow<DocumentEditSessionResponse>(r));
-
-export const endDocumentEditSession = async (
-  documentId: string,
-  sessionId: string,
-): Promise<void> => {
-  const res = await apiFetch(
-    `${API}/documents/${documentId}/edit-sessions/${sessionId}`,
-    { method: "DELETE" },
-  );
-  if (!res.ok && res.status !== 204) {
-    const text = await res.text();
-    throw new Error(`${res.status} ${res.statusText}: ${text}`);
-  }
-};
-
-export const createDocumentComment = (
-  documentId: string,
-  payload: {
-    body: string;
-    quote_text?: string | null;
-    body_sha256?: string | null;
-    anchor_start?: number | null;
-    anchor_end?: number | null;
-  },
-) =>
-  apiFetch(`${API}/documents/${documentId}/comments`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(payload),
-  }).then((r) => resolutionJsonOrThrow<DocumentCommentRead>(r));
-
-export const updateDocumentComment = (
-  documentId: string,
-  commentId: string,
-  payload: { body: string },
-) =>
-  apiFetch(
-    `${API}/documents/${encodeURIComponent(documentId)}/comments/${encodeURIComponent(
-      commentId,
-    )}`,
-    {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(payload),
-    },
-  ).then((r) => resolutionJsonOrThrow<DocumentCommentRead>(r));
-
-export const resolveDocumentComment = (documentId: string, commentId: string) =>
-  apiFetch(
-    `${API}/documents/${encodeURIComponent(documentId)}/comments/${encodeURIComponent(
-      commentId,
-    )}/resolve`,
-    { method: "POST" },
-  ).then((r) => resolutionJsonOrThrow<DocumentCommentRead>(r));
-
-export const reopenDocumentComment = (documentId: string, commentId: string) =>
-  apiFetch(
-    `${API}/documents/${encodeURIComponent(documentId)}/comments/${encodeURIComponent(
-      commentId,
-    )}/reopen`,
-    { method: "POST" },
-  ).then((r) => resolutionJsonOrThrow<DocumentCommentRead>(r));
-
-// ----- Auth + user --------------------------------------------------------
-// Auth endpoints live in `./api/auth` (sit at backend origin, NOT under
-// /api). Re-exported here so `../lib/api` consumers keep working.
-export {
-  AUTH,
-  getCurrentUser,
-  signin,
-  signout,
-  signup,
-  forgotPassword,
-  resetPassword,
-  verifyEmail,
-  requestVerifyToken,
-  updateProfile,
-} from "./api/auth";
-export type { CurrentUser, AuthError, UserProfileUpdate } from "./api/auth";
-
-// ----- Settings: API keys ------------------------------------------------
-
-export interface UserApiKeyRead {
-  provider: string;
-  last_used_at: string | null;
-  created_at: string;
-}
-
-export const listApiKeys = () =>
-  apiFetch(`${API}/settings/keys`).then((r) => jsonOrThrow<UserApiKeyRead[]>(r));
-
-export const upsertApiKey = (provider: "anthropic" | "openai", apiKey: string) =>
-  apiFetch(`${API}/settings/keys`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ provider, api_key: apiKey }),
-  }).then((r) => jsonOrThrow<UserApiKeyRead>(r));
-
-export const deleteApiKey = (provider: string) =>
-  apiFetch(`${API}/settings/keys/${encodeURIComponent(provider)}`, {
-    method: "DELETE",
-  }).then(async (r) => {
-    if (!r.ok && r.status !== 204) {
-      const text = await r.text();
-      throw new Error(`${r.status} ${r.statusText}: ${text}`);
-    }
-  });
-
-// ----- Installed-skill catalogue extensions ------------------------------
-
-// ----- Anonymisation (folded from modules/anonymisation/api.ts) ----------
-
-export type AnonymisationEngine = "presidio" | "claude" | "auto";
-
-export interface AnonymiseRequestPayload {
-  engine?: AnonymisationEngine;
-  entity_types?: string[] | null;
-  threshold?: number;
-}
-
-export interface TokenMapping {
-  token: string;
-  entity_type: string;
-  original: string;
-  occurrences: number;
-}
-
-export interface AnonymisationResult {
-  document_id: string;
-  redacted_text: string;
-  engine: string;
-  anonymised_at: string;
-  char_count: number;
-  entity_count: number;
-  tokens: TokenMapping[];
-}
-
-export interface AnonymisationSpan {
-  start: number;
-  end: number;
-  token: string;
-  original: string;
-  entity_type: string;
-}
-
-export interface MappingRead {
-  document_id: string;
-  tokens: TokenMapping[];
-  spans: AnonymisationSpan[];
-}
-
-async function anonymisationJsonOrThrow<T>(res: Response): Promise<T> {
-  if (!res.ok) {
-    let detail = "";
-    try {
-      const body = await res.json();
-      detail = typeof body?.detail === "string" ? body.detail : JSON.stringify(body);
-    } catch {
-      detail = await res.text();
-    }
-    throw new Error(`${res.status} ${res.statusText}: ${detail}`.trim());
-  }
-  return (await res.json()) as T;
-}
-
-export const anonymiseDocument = (
-  documentId: string,
-  body: AnonymiseRequestPayload = {},
-): Promise<AnonymisationResult> =>
-  apiFetch(`${API}/documents/${documentId}/anonymise`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ engine: "auto", threshold: 0.4, ...body }),
-  }).then((r) => anonymisationJsonOrThrow<AnonymisationResult>(r));
-
-export const getAnonymisation = (documentId: string): Promise<AnonymisationResult> =>
-  apiFetch(`${API}/documents/${documentId}/anonymise`).then((r) =>
-    anonymisationJsonOrThrow<AnonymisationResult>(r),
-  );
-
-export const getAnonymisationMapping = (documentId: string): Promise<MappingRead> =>
-  apiFetch(`${API}/documents/${documentId}/anonymise/mapping`).then((r) =>
-    anonymisationJsonOrThrow<MappingRead>(r),
-  );
-
-// ----- Assistant ----------------------------------------------------------
-
-export interface SuggestedAction {
-  type: "view_document" | "view_audit" | "view_chronology"
-      | "anonymise_document";
-  label: string;
-  params: Record<string, string>;
-}
-
-export interface AssistantMessage {
-  id: string;
-  role: "user" | "assistant";
-  content: string;
-  suggested_actions: SuggestedAction[];
-  model_used?: string | null;
-  created_at: string;
-}
-
-export type AssistantPostResponse = {
-  user: AssistantMessage;
-  assistant: AssistantMessage;
-};
-
-export type AssistantStreamEvent =
-  | { event: "turn.start"; data: { slug: string } }
-  | {
-      event: "context.loaded";
-      data: {
-        history_message_count: number;
-        chronology_event_count: number;
-        document_count: number;
-        tool_count: number;
-      };
-    }
-  | { event: "turn.accepted"; data: { user_message_id: string } }
-  | { event: "turn.deterministic"; data: { assistant_message_id: string; kind: string } }
-  | { event: "model.start"; data: { stage: string } }
-  | { event: "tool.start"; data: { module_id: string; capability_id: string } }
-  | {
-      event: "tool.end";
-      data: { module_id: string; capability_id: string; invocation_id: string };
-    }
-  | {
-      event: "tool.error";
-      data: { module_id: string; capability_id: string; message: string };
-    }
-  | {
-      event: "turn.end";
-      data: {
-        assistant_message_id: string;
-        tool_invocation_id: string | null;
-        tool_failed: boolean;
-      };
-    }
-  | { event: "result"; data: AssistantPostResponse }
-  | { event: "error"; data: { message: string; code?: number } };
-
-export const listAssistantMessages = (slug: string) =>
-  apiFetch(`${API}/matters/${slug}/assistant/messages`)
-    .then((r) => jsonOrThrow<AssistantMessage[]>(r));
-
-export const postAssistantMessage = (slug: string, body: { content: string; selected_document_ids?: string[] }) =>
-  apiFetch(`${API}/matters/${slug}/assistant/messages`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-  }).then((r) => jsonOrThrow<AssistantPostResponse>(r));
-
-export async function* postAssistantMessageStream(
-  slug: string,
-  body: { content: string; selected_document_ids?: string[] },
-  signal?: AbortSignal,
-): AsyncIterableIterator<AssistantStreamEvent> {
-  const resp = await apiFetch(`${API}/matters/${slug}/assistant/messages/stream`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(body),
-    signal,
-  });
-  if (!resp.ok || !resp.body) {
-    const text = await resp.text();
-    throw new Error(`${resp.status} ${resp.statusText}: ${text}`);
-  }
-  const reader = resp.body.getReader();
-  const decoder = new TextDecoder("utf-8");
-  let buffer = "";
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) return;
-    buffer += decoder.decode(value, { stream: true });
-    let sep: number;
-    while ((sep = buffer.indexOf("\n\n")) >= 0) {
-      const frame = buffer.slice(0, sep);
-      buffer = buffer.slice(sep + 2);
-      let event = "message";
-      const dataLines: string[] = [];
-      for (const line of frame.split("\n")) {
-        if (line.startsWith("event:")) event = line.slice(6).trim();
-        else if (line.startsWith("data:")) dataLines.push(line.slice(5).trim());
-      }
-      if (dataLines.length === 0) continue;
-      const data = JSON.parse(dataLines.join("\n"));
-      yield { event, data } as AssistantStreamEvent;
-    }
-  }
-}
+// Barrel for the Legalise backend API client (Fluff C1 / audit M2.1).
+//
+// The implementation lives in `lib/api/*`, split by domain. This file
+// re-exports every public symbol so existing `../lib/api` import paths
+// keep resolving unchanged — zero consumer churn. New code may import
+// from the domain modules directly.
+
+export * from "./api/_core";
+export * from "./api/auth";
+export * from "./api/matters";
+export * from "./api/documents";
+export * from "./api/assistant";
+export * from "./api/modules";
+export * from "./api/signoffs";
+export * from "./api/audit";
+export * from "./api/admin";
+export * from "./api/settings";
